Rename to `updates_pool_stream` for consistency
Was named `updates_poll_stream`.
